[#2] Added support for storing half stars in ratings.

+++ src/main/java/com/example/ratingbadge/dto/ProductDto.java
@@ -11,5 +11,5 @@ public class ProductDto {
     private UUID id;
     private String name;
     private String description;
-    private double avgStars;
+    private float avgStars;
 }
